requirements.txt
amqp==5.0.1
anyjson==0.3.3
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
billiard==3.6.3.0
blinker==1.4
celery==5.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cloud-init==20.3
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
Django==2.2.6
djangorestframework==3.10.3
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
importlib-metadata==2.0.0
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
kombu==5.0.2
language-selector==0.1
MarkupSafe==1.0
more-itertools==7.2.0
netifaces==0.10.4
oauthlib==2.0.6
PAM==0.4.2
plaid-python==7.0.0
prompt-toolkit==3.0.8
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.3
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.12
redis==3.5.3
requests==2.24.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.15.0
sqlparse==0.3.0
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
zipp==3.3.1
zope.interface==4.3.2
